gcs-ssim-comparison/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-cloud-storage==1.42.0
google-cloud-firestore==2.3.4
gunicorn==20.1.0
six==1.16.0
scikit-image
numpy
tifffile
